Fixed broken ASSERT, broken check for Lua function returns.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/branches/stone_soup@189 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/menu.cc
@@ -831,7 +831,7 @@ void column_composer::add_formatted(
         bool eol_ends_format,
         bool (*tfilt)(const std::string &))
 {
-    ASSERT(ncol >= 0 && ncol < columns.size());
+    ASSERT(ncol >= 0 && ncol < (int) columns.size());
 
     column &col = columns[ncol];
     std::vector<std::string> segs = split_string("\n", s, false, true);
